requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.5
psycopg2==2.9.3
sqlparse==0.4.2

djangorestframework~=3.13.1